feat: add custom markers for map components
