+++ roamance/src/components/maps/markers/index.ts
@@ -1,0 +1,8 @@
+'use client';
+
+export * from './CustomStartPointMarker';
+export * from './DestinationMarker';
+export * from './POIMarker';
+export * from './SearchPinMarker';
+export * from './UserLocationMarker';
+export * from './WaypointMarker';
